refactor(server-remote-cloudflare): inline worker.ts into app.ts and move env.ts to tooling/
worker.ts was a pure barrel re-exporting from app.ts and yjs-room.ts.
Since wrangler supports any file as entry point, the indirection was
unnecessary — app.ts now directly exports the ExportedHandler default
and the YjsRoom Durable Object class.

env.ts (CLI-only utility for drizzle-kit and better-auth CLI) moved
from src/ to tooling/ to clarify it is not part of the worker runtime.
