+++ packages/server-remote-cloudflare/better-auth.config.ts
@@ -4,7 +4,7 @@
  * Run via:
  *   bun run auth:generate  — generate Drizzle schema from Better Auth tables
  *
- * Loads `.dev.vars` via `env.ts`.
+ * Loads `.dev.vars` via `tooling/env.ts`.
  * Schema-affecting options (basePath, plugins, emailAndPassword) come from
  * the shared `better-auth-base` — the single source of truth.
  */
@@ -14,7 +14,7 @@ import { betterAuth } from 'better-auth';
 import { drizzleAdapter } from 'better-auth/adapters/drizzle';
 import { drizzle } from 'drizzle-orm/postgres-js';
 import postgres from 'postgres';
-import { env } from './src/env';
+import { env } from './tooling/env';
 
 const sql = postgres(env.DATABASE_URL);
 const db = drizzle(sql);
