--- packages/server-remote-cloudflare/src/worker.ts
@@ -1,2 +1,0 @@
-export { default } from './app';
-export { YjsRoom } from './yjs-room';
