+++ packages/server-remote-cloudflare/wrangler.toml
@@ -1,5 +1,5 @@
 name = "epicenter-hub"
-main = "src/worker.ts"
+main = "src/app.ts"
 compatibility_date = "2025-01-01"
 compatibility_flags = ["nodejs_compat"]
